llamaindex/requirements.txt
llama-index-core>=0.11.0,<1.0.0
llama-index-llms-openai>=0.2.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
